Update VirtualJaguar to latest
Sound now works

+++ src/crc32.cpp
@@ -14,7 +14,6 @@
 //
 
 #include "crc32.h"
-#include <iostream>
 
 static unsigned long crctable[256] =
 {
@@ -59,6 +58,6 @@ int crc32_calcCheckSum(unsigned char * data, unsigned int length)
 
 	for(unsigned int i=0; i<length; i++)
 		crc = crctable[(crc ^ *data++) & 0xFF] ^ (crc >> 8);
-    
-	return (int)crc ^ 0xFFFFFFFF;
+
+	return crc ^ 0xFFFFFFFF;
 }
